requirements.txt
fastapi 
uvicorn 
transformers 
gitpython 
python-dotenv
jsonpatch
huggingface_hub
torch 
tensorflow 
flax
